==> 2a Extracting sign distributions.ipynb <==
# Extracting sign distributions

This notbook contains code for:
- importing necessary libraries
- loading stored WTs
- creating random sparse networks
- Extracting sign distributions from WTs
- Extracting sign distributions from non-WTs
- saving extracted distributions in files

### Requirements:

# importing necessary libraries

from cnn_architecture import CNN2Model
from utils import *
from load_datasets import load_and_prep_dataset

import tensorflow_datasets as tfds
import pandas as pd
import tensorflow as tf
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from tqdm import tqdm
import random

# extract sign distribution from one layer

def sign_distribution_layer(this_layer, next_layer):
    '''
    columns: 
    ["prune_rate_in", "prune_rate_out", "pos_in", "pos_out", "neg_in", "neg_out", "sign_rate_in", "sign_rate_out"]
    '''
    sign_distr = pd.DataFrame()
    unconnected_neurons = {"no_incomming": 0 , "no_outgoing" : 0, "no_total_conncetions" : 0}
    for this_neuron, next_neurons, i in zip(this_layer.T, next_layer, range(np.shape(this_layer)[0])):
        
        this_unpruned = this_neuron[this_neuron != 0.0]
        next_unpruned = next_neurons[next_neurons != 0.0]
        
        # only include neurons that receive and propagate input
        if len(this_unpruned)*len(next_unpruned) != 0:
            
            dic = {} 
            dic["pos_in"] = len(this_unpruned[this_unpruned>0])
            dic["pos_out"] = len(next_unpruned[next_unpruned>0])
            dic["neg_in"] = len(this_unpruned[this_unpruned<0])
            dic["neg_out"] = len(next_unpruned[next_unpruned<0])
            dic["prune_rate_in"] = 1 - (len(this_unpruned)/len(this_neuron))
            dic["prune_rate_out"] = 1 - (len(next_unpruned)/len(next_neurons))
            dic["sign_rate_in"] = dic["pos_in"]/len(this_unpruned)   
            dic["sign_rate_out"] = dic["pos_out"]/len(next_unpruned)                               

            df = pd.DataFrame(data = dic, index = [i])
            sign_distr = pd.concat([sign_distr, df], axis=0)
        
        # store information about inconnected neurons
        else :
            
            if len(this_unpruned) == 0:
                if len(next_unpruned) ==0:
                    unconnected_neurons["no_total_conncetions"] = unconnected_neurons["no_total_conncetions"] + 1
                else:
                    unconnected_neurons["no_incomming"] = unconnected_neurons["no_incomming"] + 1
            else:
                unconnected_neurons["no_outgoing"] = unconnected_neurons["no_outgoing"] + 1
        
    return sign_distr, unconnected_neurons

# extract sign distribution form two layers

def sign_distribution_layers(layer1, layer2, layer3):
    
    sign_distr1, unconnected_neurons1 = sign_distribution_layer(layer1, layer2)
    sign_distr2, unconnected_neurons2 = sign_distribution_layer(layer2, layer3)
    sign_distr1["layer"] = "dense1"
    sign_distr2["layer"] = "dense2"
    sign_distr_both = pd.concat([sign_distr1, sign_distr2], axis=0)
    
    unconnected_neurons1["layer"] = "dense1"
    unconnected_neurons2["layer"] = "dense2"

    return sign_distr_both, unconnected_neurons1, unconnected_neurons2

## Extracting sign distributions and storing them as files:

wts_per_dataset = 15

# extract sign distributions

def get_sign_distr(dataset, pruning_name, n):

    # collect statistics of unconnected neurons
    unconnected_statistics = pd.DataFrame()
    
    # make a model to load the weights into
    train_dataset, test_dataset = load_and_prep_dataset(dataset, batch_size=60, shuffle_size=512)
    model = CNN2Model()
    model(list(train_dataset)[0][0])
    for i in tqdm(range(n), leave=False, desc=f"sign_distributions of {dataset}"):
        
        # get WT weights
        model.load_weights(f"1b WTs/WT_{dataset}_{pruning_name}_{i}.h5")    
        weights_wt = model.get_weights()
        # get sign distribution
        sign_distr_wt, unconnected_neurons1, unconnected_neurons2 = sign_distribution_layers(weights_wt[4], weights_wt[6], weights_wt[8])
        # store sign distribution
        sign_distr_wt.to_csv(f'2b Sign distributions/{dataset}_{pruning_name}_{i}_sign_distr.csv', index=False)
        
        
        #collect unconnected neurons statistics
        unconnected_neurons1["model"] = f"WT_{dataset}_{pruning_name}_{i}"
        unconnected_neurons2["model"] = f"WT_{dataset}_{pruning_name}_{i}"
        df1 = pd.DataFrame(data = unconnected_neurons1, index = [i])
        df2 = pd.DataFrame(data = unconnected_neurons2, index = [i])
        unconnected_statistics = pd.concat([unconnected_statistics, df1, df2], axis=0)
        
    unconnected_statistics.to_csv(f'2b Sign distributions/{dataset}_{pruning_name}_unconnected_statistic.csv', index=False)

# store sign distributions of Wts

#for dataset in ["SVHN","CINIC","CIFAR"]:
#    get_sign_distr(dataset,"IMP",wts_per_dataset)

### Generating randomly pruned sign distributions

# implement random pruning of a weight matrix according to a pruning rate

def random_pruning(model,pruning_rates):
    
    weights = model.get_weights()
    new_weights = []
    
    for layer, p_rate in zip(weights[::2], pruning_rates):
        shape = np.shape(layer)
        layer = layer.flatten()
        number_to_prune = int(len(layer) * p_rate)
        pruning_indexi = random.sample(range(len(layer)),number_to_prune)
        layer[pruning_indexi] = 0.0
        layer = np.reshape(layer,shape)
        new_weights.append(layer)
        
    weights[::2] = new_weights
    model.set_weights(weights)
    weights = model.get_weights()
    
    return model

def make_random_sign_distr(n):

    # get pruning rates from imp model
    train_dataset, test_dataset = load_and_prep_dataset("CIFAR",batch_size=60, shuffle_size=512)
    model_cifar_imp = CNN2Model()
    model_cifar_imp(list(train_dataset)[0][0])
    model_cifar_imp.load_weights("1b WTs/h5_format/WT_CIFAR_IMP_0.h5")
    pruning_rates = get_pruning_rates(model_cifar_imp.get_weights()[::2])

    unconnected_statistics = pd.DataFrame()

    # generate random sparse networks
    for i in range(n):
        
        # randomly prune a model
        random_model = CNN2Model()
        random_model(list(train_dataset)[0][0])
        random_model = random_pruning(random_model,pruning_rates)
        random_model.save_weights(f"./1b WTs/h5_format/RSN_{i}.h5", overwrite=True)
        weights_rsn = random_model.get_weights()
        
        # store sign distributions of random sparse networks
        sign_distr_random, unconnected_neurons1, unconnected_neurons2  = sign_distribution_layers(weights_rsn[4], weights_rsn[6], weights_rsn[8])
        sign_distr_random.to_csv(f"2b Sign distributions/RSN_{i}_sign_distr.csv", index=False)
        
        #collect unconnected neurons statistics
        unconnected_neurons1["model"] = f"RSN_{i}"
        unconnected_neurons2["model"] = f"RSN_{i}"
        df1 = pd.DataFrame(data = unconnected_neurons1, index = [i])
        df2 = pd.DataFrame(data = unconnected_neurons2, index = [i])
        unconnected_statistics = pd.concat([unconnected_statistics, df1, df2], axis=0)

    # store unconnected statistics
    unconnected_statistics.to_csv(f'2b Sign distributions/RSN_unconnected_statistic.csv', index=False)

#make_random_sign_distr(wts_per_dataset)

## Get Pruning rates

train_dataset, test_dataset = load_and_prep_dataset("CIFAR", batch_size=60, shuffle_size=512)
model_cifar_imp = CNN2Model()
model_cifar_imp(list(train_dataset)[0][0])
model_cifar_imp.load_weights("1b WTs/h5_format/WT_CIFAR_IMP_0.h5")

# return the pruning rates of the conv and dense layers
print("conv rate (without biases): ", get_pruning_rate(model_cifar_imp.get_weights()[0:4:2]))
print("dense rate (without biases): ", get_pruning_rate(model_cifar_imp.get_weights()[4::2]))
print("conv rate (with biases): ", get_pruning_rate(model_cifar_imp.get_weights()[0:4:]))
print("dense rate (with biases): ", get_pruning_rate(model_cifar_imp.get_weights()[4::]))
print("pruning rates without biases each layer: ", get_pruning_rates(model_cifar_imp.get_weights()[::2]))
print("pruning rates with biases each layer: ")
print("    ", get_pruning_rate(model_cifar_imp.get_weights()[0:2]))
print("    ", get_pruning_rate(model_cifar_imp.get_weights()[2:4]))
print("    ", get_pruning_rate(model_cifar_imp.get_weights()[4:6]))
print("    ", get_pruning_rate(model_cifar_imp.get_weights()[6:8]))
print("    ", get_pruning_rate(model_cifar_imp.get_weights()[8:10]))

print("pruning rate total with biases: ", get_pruning_rate(model_cifar_imp.get_weights()))
print("pruning rate total without biases: ", get_pruning_rate(model_cifar_imp.get_weights()[::2]))

# return the parameter counts of the conv and dense layers
print("unpruned parameter counts layer types")
print("conv rate (without biases): ", get_unpruned_params_count(model_cifar_imp.get_weights()[0:4:2]))
print("dense rate (without biases): ", get_unpruned_params_count(model_cifar_imp.get_weights()[4::2]))
print("conv rate (with biases): ", get_unpruned_params_count(model_cifar_imp.get_weights()[0:4:]))
print("dense rate (with biases): ", get_unpruned_params_count(model_cifar_imp.get_weights()[4::]))
print("")
print("unpruned parameter counts individual layers")
print("without biases each layer: ", get_unpruned_params_counts(model_cifar_imp.get_weights()[::2]))
print("with biases each layer: ")
print("    ", get_unpruned_params_count(model_cifar_imp.get_weights()[0:2]))
print("    ", get_unpruned_params_count(model_cifar_imp.get_weights()[2:4]))
print("    ", get_unpruned_params_count(model_cifar_imp.get_weights()[4:6]))
print("    ", get_unpruned_params_count(model_cifar_imp.get_weights()[6:8]))
print("    ", get_unpruned_params_count(model_cifar_imp.get_weights()[8:10]))
print("")
print("unpruned parameter counts total with biases: ", get_unpruned_params_count(model_cifar_imp.get_weights()))
print("unpruned parameter counts total without biases: ", get_unpruned_params_count(model_cifar_imp.get_weights()[::2]))

# return the parameter counts of the conv and dense layers
print("parameter counts layer types")
print("conv rate (without biases): ", get_params_count(model_cifar_imp.get_weights()[0:4:2]))
print("dense rate (without biases): ", get_params_count(model_cifar_imp.get_weights()[4::2]))
print("conv rate (with biases): ", get_params_count(model_cifar_imp.get_weights()[0:4:]))
print("dense rate (with biases): ", get_params_count(model_cifar_imp.get_weights()[4::]))
print("")
print("parameter counts individual layers")
print("without biases each layer: ", get_params_counts(model_cifar_imp.get_weights()[::2]))
print("with biases each layer: ")
print("    ", get_params_count(model_cifar_imp.get_weights()[0:2]))
print("    ", get_params_count(model_cifar_imp.get_weights()[2:4]))
print("    ", get_params_count(model_cifar_imp.get_weights()[4:6]))
print("    ", get_params_count(model_cifar_imp.get_weights()[6:8]))
print("    ", get_params_count(model_cifar_imp.get_weights()[8:10]))
print("")
print("unpruned parameter counts total with biases: ", get_params_count(model_cifar_imp.get_weights()))
print("unpruned parameter counts total without biases: ", get_params_count(model_cifar_imp.get_weights()[::2]))
